rename icdar txt parser
