+++ mmocr/datasets/preparers/parsers/icdar_txt_parser.py
